backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
aniso8601==10.0.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
attrs==25.4.0
billiard==4.2.4
blinker==1.9.0
cachelib==0.13.0
celery==5.6.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.1
Flask==3.1.2
Flask-Caching==2.3.1
flask-cors==6.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security==5.7.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flower==2.0.1
frozenlist==1.8.0
greenlet==3.3.0
humanize==4.14.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.6.1
libpass==1.9.3
MarkupSafe==3.0.3
multidict==6.7.0
packaging==25.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
redis==7.1.0
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.44
tornado==6.5.2
twilio==9.8.8
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.6.0
vine==5.1.0
wcwidth==0.2.14
Werkzeug==3.1.4
WTForms==3.2.1
yarl==1.22.0
